requirements.txt
python-decouple==3.4
requests==2.26.0